+++ platformio.ini
@@ -17,3 +17,5 @@
 platform = atmelsam
 board = adafruit_feather_m0
 framework = arduino
+lib_deps =
+  https://github.com/arduino-libraries/ArduinoHttpClient.git
