+++ src/FeatherClock.cpp
@@ -255,75 +255,6 @@ void sendNTPpacket(IPAddress &address) {
   result = Udp.endPacket();
 }
 
-// Connect to the server and read the alarm time
-// 0000 means no alarm
-void getAlarmTime(String url) {
-  WiFiClient client;
-  const int httpPort = 80;
-  int endServer = -1;   // locates the third slash in the URL
-  int hostStart = -1;   // locates the start of the hostname, after the leading protocol info (e.g. http://)
-  String server;
-  String host;
-  String filePart;      // Holds the part of the URL after the server name
-  String payload;       // string read from the server
-
-
-  // ASSUMPTION: The third "/" is the end of the server name. There must be a file path after that, e.g. foo.php
-  // This is based on http{s}://server.foo/something.php format
-  /*endServer = url.indexOf('/');                // First /
-  endServer = url.indexOf('/', endServer+1);    // Second /
-  hostStart = endServer;    // So we can skip the protocol to grab just the hostname for the request
-  endServer = url.indexOf('/', endServer+1);    // Third /
-  // At this point we either know the locations of the 3rd slash or endServer = -1 which means we didn't find it
-  if ((endServer == -1) || (hostStart == -1)) {
-    Serial.print("Failed to parse server portion of URL"); Serial.println(url);
-    return;
-  }
-  Serial.print("hostStart = "); Serial.println(hostStart);
-  Serial.print("endServer = "); Serial.println(endServer);
-  server = url.substring(hostStart+1, endServer);
-  host = url.substring(hostStart+1, endServer);
-  filePart = url.substring(endServer+1);*/
-
-  //url = "www.arduino.cc/latest.txt";
-
-  server = url.substring(0, url.indexOf('/'));
-  filePart = url.substring(url.indexOf('/'));
-
-  Serial.print("Server: " + server);
-  Serial.println(", Filepart: " + filePart);
-  //Serial.print(F("Parsed host name of: ")); Serial.println(host);
-
-  //char server2[] = "www.arduino.cc";
-  if (client.connect(server.c_str(), 80)) {
-    Serial.println(F("Connected to server"));
-    client.println("GET " + filePart + " HTTP/1.1");
-    client.println("Host: " + server);
-    client.println("User-Agent: ArduinoWiFi/1.1");
-    client.println("Connection: close");
-    client.println();
-
-    char c;
-    Serial.println("Wait for client data");
-    while (!client.available()) {
-      //char c = client.read();
-      //Serial.write(c);
-    }
-    while(client.available()) {
-      c = client.read();
-      Serial.write(c);
-      payload += c;
-    }
-    Serial.println("<end>");
-    client.stop();
-
-    alarmHour = atoi(payload.substring(0,2).c_str());
-    alarmMinute = atoi(payload.substring(2,4).c_str());
-  } else {
-    Serial.println("Failed to connect to server");
-  }
-}
-
 boolean alarmTime() {
 
   // Sound continuously if debugging
